# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "contact": {
      "name": "Ferme des animaux - IIM",
      "url": "https://github.com/alyssialopr/FermeDesAnimauxJava"
    },
    "description": "Gestion d'une ferme : des **eleveurs** achetent, vendent, nourrissent, soignent et promenent des **vaches** et des **poules**.\n\n### Principes\n\n- un animal peut arriver a la ferme sans proprietaire, puis etre achete ;\n- un eleveur ne peut agir que sur ses propres animaux, sinon la reponse est `409` avec le message *Cet animal ne vous appartient pas* ;\n- un animal `VENDU`, `MORT` ou `DISPARU` n'accepte plus aucune action ;\n- chaque action renvoie le message metier correspondant ainsi que l'etat de l'animal apres l'action.\n\n### Erreurs\n\nToutes les erreurs suivent la RFC 7807 (`application/problem+json`) : `title`, `status`, `detail`, `instance`, `horodatage`, et `champs` pour les erreurs de validation.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "API Ferme des animaux",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Environnement local (docker compose)"
    }
  ],
...
  "paths": {
...
    "/api/eleveurs": {
      "get": {
        "description": "Version resumee : le champ `animaux` est null, seul `nombreAnimaux` est renseigne.",
        "operationId": "recupereEleveurs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EleveurResponse"
                  }
                }
              }
            },
            "description": "Eleveurs de la ferme, tries par prenom"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/eleveurs",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "prenom": "le prenom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Eleveur introuvable : 42",
                      "instance": "/api/eleveurs/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Eleveur ou animal inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Animal d'un autre eleveur": {
                    "description": "Animal d'un autre eleveur",
                    "value": {
                      "title": "Animal non possede",
                      "status": 409,
                      "detail": "Cet animal ne vous appartient pas",
                      "instance": "/api/eleveurs/2/animaux/1/repas",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  },
                  "Animal indisponible": {
                    "description": "Animal indisponible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/eleveurs/1/animaux/2/repas",
                      "horodatage": "2026-09-18T07:27:13.527Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal appartenant a un autre eleveur, animal vendu/mort/disparu, prenom deja pris, ou suppression d'un eleveur dont le troupeau n'est pas vide"
          }
        },
        "summary": "Lister les eleveurs",
        "tags": [
          "Eleveurs"
        ]
      },
      "post": {
        "description": "Le prenom doit etre unique a la ferme. La reponse contient la **cle d'acces** de l'eleveur : c'est la seule fois qu'elle est renvoyee, le serveur n'en garde qu'une empreinte BCrypt. Elle sert ensuite a s'authentifier (`Authorization: Bearer <id>.<cle>`).",
        "operationId": "creerEleveur",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreerEleveurRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EleveurCreeResponse"
                }
              }
            },
            "description": "Eleveur cree, avec sa cle d'acces"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/eleveurs",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "prenom": "le prenom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Eleveur introuvable : 42",
                      "instance": "/api/eleveurs/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Eleveur ou animal inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Animal d'un autre eleveur": {
                    "description": "Animal d'un autre eleveur",
                    "value": {
                      "title": "Animal non possede",
                      "status": 409,
                      "detail": "Cet animal ne vous appartient pas",
                      "instance": "/api/eleveurs/2/animaux/1/repas",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  },
                  "Animal indisponible": {
                    "description": "Animal indisponible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/eleveurs/1/animaux/2/repas",
                      "horodatage": "2026-09-18T07:27:13.527Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal appartenant a un autre eleveur, animal vendu/mort/disparu, prenom deja pris, ou suppression d'un eleveur dont le troupeau n'est pas vide"
          }
        },
        "summary": "Creer un eleveur",
        "tags": [
          "Eleveurs"
        ]
      }
    },
    "/api/eleveurs/classement": {
      "get": {
        "description": "La fortune, c'est l'argent en caisse plus la valeur de revente du troupeau.",
        "operationId": "recupereClassement",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassementResponse"
                  }
                }
              }
            },
            "description": "Classement, du plus riche au plus pauvre"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/eleveurs",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "prenom": "le prenom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Eleveur introuvable : 42",
                      "instance": "/api/eleveurs/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Eleveur ou animal inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Animal d'un autre eleveur": {
                    "description": "Animal d'un autre eleveur",
                    "value": {
                      "title": "Animal non possede",
                      "status": 409,
                      "detail": "Cet animal ne vous appartient pas",
                      "instance": "/api/eleveurs/2/animaux/1/repas",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  },
                  "Animal indisponible": {
                    "description": "Animal indisponible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/eleveurs/1/animaux/2/repas",
                      "horodatage": "2026-09-18T07:27:13.527Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal appartenant a un autre eleveur, animal vendu/mort/disparu, prenom deja pris, ou suppression d'un eleveur dont le troupeau n'est pas vide"
          }
        },
        "summary": "Classer les eleveurs par fortune",
        "tags": [
          "Eleveurs"
        ]
      }
    },
    "/api/eleveurs/{eleveurId}/animaux/{animalId}/achat": {
      "post": {
        "description": "Rattache l'animal a l'eleveur et le remet a l'etat `LIBRE`. Refuse si l'animal appartient deja a quelqu'un d'autre, ou s'il est mort ou disparu.",
        "operationId": "acheter",
        "parameters": [
          {
            "description": "Identifiant de l'acheteur",
            "example": 2,
            "in": "path",
            "name": "eleveurId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "description": "Identifiant de l'animal",
            "example": 5,
            "in": "path",
            "name": "animalId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "message": "killian a achete la poule caramel.",
                  "animal": {
                    "id": 5,
                    "etat": "LIBRE"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Animal achete"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/eleveurs",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "prenom": "le prenom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Eleveur introuvable : 42",
                      "instance": "/api/eleveurs/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Eleveur ou animal inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Animal d'un autre eleveur": {
                    "description": "Animal d'un autre eleveur",
                    "value": {
                      "title": "Animal non possede",
                      "status": 409,
                      "detail": "Cet animal ne vous appartient pas",
                      "instance": "/api/eleveurs/2/animaux/1/repas",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  },
                  "Animal indisponible": {
                    "description": "Animal indisponible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/eleveurs/1/animaux/2/repas",
                      "horodatage": "2026-09-18T07:27:13.527Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal appartenant a un autre eleveur, animal vendu/mort/disparu, prenom deja pris, ou suppression d'un eleveur dont le troupeau n'est pas vide"
          }
        },
        "summary": "Acheter un animal",
        "tags": [
          "Eleveurs"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreerEleveurRequest": {
        "type": "object",
        "description": "Nouvel eleveur de la ferme",
        "properties": {
          "prenom": {
            "type": "string",
            "description": "Prenom de l'eleveur, unique a la ferme (casse ignoree)",
            "example": "alyssia",
            "maxLength": 40,
            "minLength": 0,
            "pattern": "[\\p{L} '\\-]+"
          }
        },
        "required": [
          "prenom"
        ]
      },
      "EleveurCreeResponse": {
        "type": "object",
        "description": "Eleveur tout juste cree, avec sa cle d'acces. C'est la seule et unique fois que la cle est renvoyee : le serveur n'en garde qu'une empreinte.",
        "properties": {
          "cle": {
            "type": "string",
            "description": "Cle d'acces, a conserver : elle autorise toutes les actions au nom de cet eleveur (en-tete Authorization: Bearer <id>.<cle>)",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "eleveur": {
            "$ref": "#/components/schemas/EleveurResponse",
            "description": "L'eleveur cree"
          }
        }
      },
      "EleveurResponse": {
        "type": "object",
        "description": "Eleveur, son porte-monnaie et, selon la route, son troupeau",
        "properties": {
          "animaux": {
            "type": "array",
            "description": "Troupeau detaille. Renseigne sur la fiche d'un eleveur, null dans la liste des eleveurs",
            "items": {
              "$ref": "#/components/schemas/AnimalResponse"
            }
          },
          "fortune": {
            "type": "number",
            "description": "Solde plus la valeur de revente du troupeau",
            "example": 662.6
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Identifiant de l'eleveur",
            "example": 1
          },
          "nombreAnimaux": {
            "type": "integer",
            "format": "int32",
            "description": "Taille du troupeau",
            "example": 2
          },
          "prenom": {
            "type": "string",
            "description": "Prenom de l'eleveur",
            "example": "alyssia"
          },
          "solde": {
            "type": "number",
            "description": "Argent disponible, en euros",
            "example": 248.6
          }
        }
      },
      "Erreur": {
        "type": "object",
        "description": "Erreur normalisee au format RFC 7807 (application/problem+json)",
        "properties": {
          "champs": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Detail par champ, uniquement pour les erreurs de validation (400)",
            "example": {
              "nom": "le nom est obligatoire"
            }
          },
          "detail": {
            "type": "string",
            "description": "Message explicatif",
            "example": "Cet animal ne vous appartient pas"
          },
          "horodatage": {
            "type": "string",
            "format": "date-time",
            "description": "Date de l'erreur",
            "example": "2026-09-18T07:27:13.445Z"
          },
          "instance": {
            "type": "string",
            "description": "Chemin de la requete fautive",
            "example": "/api/eleveurs/2/animaux/1/repas"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "Code HTTP",
            "example": 409
          },
          "title": {
            "type": "string",
            "description": "Intitule court de l'erreur",
            "example": "Animal non possede"
          }
        }
      },
...
    }
  }
}